Do logic at compile time
