+++ tests/test_wire/test_967.py
@@ -24,10 +24,8 @@ def test_missing_generator_paren(caplog):
 
     with pytest.raises(MagmaCompileException) as e:
         m.compile("build/Bar", Bar(4))
-    assert str(e.value) == "Found circuit with errors: Bar"
-    assert (str(caplog.records[0].msg) ==
-            "Cannot wire Bar.a to Foo.y because they are not from the same"
-            " definition")
+    assert str(e.value) == ("Cannot wire Bar.a to Foo.y because they are not "
+                            "from the same definition")
 
 
 def test_bad_temp(caplog):
@@ -45,7 +43,22 @@ def test_bad_temp(caplog):
 
     with pytest.raises(MagmaCompileException) as e:
         m.compile("build/Bar", Bar)
-    assert str(e.value) == "Found circuit with errors: Bar"
-    assert (str(caplog.records[0].msg) ==
-            "Cannot wire Foo.x to Bar.y because they are not from the same"
-            " definition")
+    assert str(e.value) == ("Cannot wire Foo.x to Bar.y because they are not "
+                            "from the same definition")
+
+
+def test_bad_temp2(caplog):
+    class Foo(m.Circuit):
+        io = m.IO(x=m.In(m.Bits[8]), y=m.Out(m.Bits[8]))
+        z = m.Bits[8]()
+        io.y @= z
+
+    class Bar(m.Circuit):
+        io = m.IO(x=m.In(m.Bits[8]), y=m.Out(m.Bits[8]))
+        Foo.z @= io.x
+        io.y @= Foo.z
+
+    with pytest.raises(MagmaCompileException) as e:
+        m.compile("build/Foo", Foo)
+    assert str(e.value) == ("Cannot wire Bar.x to Foo.y because they are not "
+                            "from the same definition")
